Suffix model concerns with 'able'
