+++ app/models/edm/place.rb
@@ -7,7 +7,7 @@ module EDM
     include Mongoid::Uuid
     include AutocompletableModel
     include Blankness::Mongoid
-    include RDF::Model
+    include RDF::Graphable
 
     has_one :dcterms_spatial_place_for,
             class_name: 'EDM::ProvidedCHO', inverse_of: :dcterms_spatial_places
