+++ spec/models/edm/place_spec.rb
@@ -7,6 +7,6 @@ RSpec.describe EDM::Place do
     it { is_expected.to include(Mongoid::Timestamps) }
     it { is_expected.to include(Mongoid::Uuid) }
     it { is_expected.to include(Blankness::Mongoid) }
-    it { is_expected.to include(RDF::Model) }
+    it { is_expected.to include(RDF::Graphable) }
   end
 end
