+++ spec/models/edm/web_resource_spec.rb
@@ -10,7 +10,7 @@ RSpec.describe EDM::WebResource do
     it { is_expected.to include(Mongoid::Timestamps) }
     it { is_expected.to include(Mongoid::Uuid) }
     it { is_expected.to include(Blankness::Mongoid) }
-    it { is_expected.to include(RDF::Model) }
+    it { is_expected.to include(RDF::Graphable) }
 
     it { is_expected.to reject_if_blank(:dc_creator_agent) }
   end
